+++ packages/select-text/configure/src/index.js
@@ -7,9 +7,9 @@ import {
   ModelUpdatedEvent
 } from '@pie-framework/pie-configure-events';
 import defaults from 'lodash/defaults';
-import defaultValues from './defaultConfigure';
+import defaultValues from './defaultConfiguration';
 
-const prepareCustomizationObject = (model) => {
+const prepareCustomizationObject = model => {
   return {
     ...model,
     configure: defaults(model.configure, defaultValues)