+++ packages/select-text/configure/CHANGELOG.md
@@ -3,6 +3,43 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.2.2](https://github.com/pie-framework/pie-elements/compare/@pie-element/select-text-configure@2.2.1...@pie-element/select-text-configure@2.2.2) (2018-11-12)
+
+
+### Bug Fixes
+
+* **dependencies-add:** added dependencies where needed ([23df697](https://github.com/pie-framework/pie-elements/commit/23df697))
+* conflict ([b854c6b](https://github.com/pie-framework/pie-elements/commit/b854c6b))
+
+
+
+
+
+<a name="2.2.1"></a>
+## [2.2.1](https://github.com/pie-framework/pie-elements/compare/@pie-element/select-text-configure@2.2.0...@pie-element/select-text-configure@2.2.1) (2018-11-08)
+
+
+### Bug Fixes
+
+* fix defaultConfiguration import error ([a092493](https://github.com/pie-framework/pie-elements/commit/a092493))
+
+
+
+
+
+<a name="2.2.0"></a>
+# [2.2.0](https://github.com/pie-framework/pie-elements/compare/@pie-element/select-text-configure@2.1.0...@pie-element/select-text-configure@2.2.0) (2018-11-05)
+
+
+### Features
+
+* **select-text-configure-customization:** created an exportable module for the defaultValues and added some tests for customization ([e047522](https://github.com/pie-framework/pie-elements/commit/e047522))
+* **select-text-configure-customization:** made it possible to customize the elements from the config design, updated tests ([9fb0fd4](https://github.com/pie-framework/pie-elements/commit/9fb0fd4))
+
+
+
+
+
 <a name="2.1.0"></a>
 # [2.1.0](https://github.com/pie-framework/pie-elements/compare/@pie-element/select-text-configure@2.0.0...@pie-element/select-text-configure@2.1.0) (2018-10-31)